restructure project; remove old dashboard components
