+++ src/components/AppsComponents/AppInstallCard/AppInstallCard.tsx
