+++ src/components/HomeComponents/BitcoinCard/BitcoinCard.tsx
