+++ src/components/HomeComponents/TransactionList/TransactionDetailModal/LNDetails/LNDetails.tsx
